Finish non-recursive FFT
In this verision, I add linux-list as the base of task queue. In
the process of fourier transform, the data will be splited by an
divisor again and again which divide the problem to several tiny
sub-problems. The problem and its sub-problems will be linked on
the task queue. In each iteration, the queue pop up an task and
do something on it depending its type either DIVIDE or MERGE. If
the task's type is DIVIDE, the task will be divided to several
sub-probems or finish the task if it's scale is tiny. If its type
is MERGE, it merges the subproblems' result.