Optimize process of init omega matrix
In previous version, the function directly compute each matrix
element, but it waste a lot of time because there is a relationship
between each element which is conjugate and the phase of exponential.
For example, W_5^1 = (W_5^4)*. For another example, W_5^6 = W_5^1.
I use these relations to reduce the computation of cos and sin
function