Finish inverse fast fourier transform
In this commit, I integrate ifft into fft because basically they
are the same algorithm only depending on the omega matrix and the
the process of merge. For these two factors, I add a function
pointer array and a function pointer into the function, but for easily
testing, the original FFT declartion is remained. The users would'nt
change the code for this modification.

Actually, readability is trade-off of efficiency, but I valued
readability more than efficiency. Therefore, in this commit, you may
probabily notice the code is a little slower than last version.